+++ Braintree/src/test/java/com/braintreepayments/api/PayPalRequestUnitTest.java
@@ -139,6 +139,7 @@ public class PayPalRequestUnitTest {
                 .userAction(PayPalRequest.USER_ACTION_COMMIT)
                 .displayName("Display Name")
                 .offerCredit(true)
+                .offerPayLater(true)
                 .merchantAccountId("merchant_account_id");
 
         ArrayList<PayPalLineItem> lineItems = new ArrayList<PayPalLineItem>();
@@ -164,6 +165,7 @@ public class PayPalRequestUnitTest {
         assertEquals(PayPalRequest.USER_ACTION_COMMIT, result.getUserAction());
         assertEquals("Display Name", result.getDisplayName());
         assertTrue(result.shouldOfferCredit());
+        assertTrue(result.shouldOfferPayLater());
         assertEquals("merchant_account_id", result.getMerchantAccountId());
         assertEquals(1, result.getLineItems().size());
         assertEquals("An Item", result.getLineItems().get(0).getName());